Attempt to fix frequent MiniPlayerFragment ANRs seen in production
For background on the issue, and why this change may help, see
https://github.com/androidx/media/issues/890

The basic idea here is to reduce the rate at which we create and release
MediaControllers in MiniPlayerFragment, by moving the creation and
releasing to fragment lifecycle callbacks that are invoked less
frequently than `onResume()` and `onPause()`. It's not completely clear
that this will definitely solve the problems we're seeing in production,
but it seems like a worthwhile experiment.